+++ src/core/execution/runners/sandboxRunner.ts
@@ -41,6 +41,7 @@ export abstract class SandboxRunner<TContext = unknown> {
   protected getContextInitParams(descriptor: SandboxExecutionDescriptor): MacroContextInitParams {
     return {
       disposables: descriptor.macroDisposables,
+      extensionContext: this.context.extensionContext,
       log: new MacroLogOutputChannel(descriptor.id as any, this.context),
       executionId: descriptor.id as any,
       startup: descriptor.startup,