+++ src/macroPseudoterminal.ts
@@ -44,6 +44,7 @@ export class MacroPseudoterminal implements vscode.Pseudoterminal {
 
     this.macroInitParams = {
       disposables: [],
+      extensionContext: this.context.extensionContext,
       log: new MacroLogOutputChannel(runId, context),
       executionId: runId,
       token: this.cts.token,